+++ src/lib/utils/repo/account.repo.ts
@@ -249,3 +249,23 @@ export async function findUniqueAccountByCodeInPrisma(code: string, companyId?: 
   });
   return account;
 }
+
+export async function fuzzySearchAccountByName(name: string) {
+  let account: Account | null = null;
+
+  try {
+    const accounts: Account[] = await prisma.$queryRaw`
+      SELECT * FROM public."account"
+      WHERE for_user = true
+      ORDER BY SIMILARITY(name, ${name}) DESC
+      LIMIT 1;
+    `;
+    [account] = accounts;
+  } catch (error) {
+    // Deprecated: （ 20240619 - Murky）Debugging purpose
+    // eslint-disable-next-line no-console
+    console.error(error);
+  }
+
+  return account;
+}
