+++ src/pages/api/v1/company/[companyId]/account/[accountId]/index.ts
@@ -14,7 +14,6 @@ import { getSession } from '@/lib/utils/session';
 import { AuthFunctionsKeys } from '@/interfaces/auth';
 
 function formatParams(companyId: unknown, accountId: string | string[] | undefined) {
-  // ToDo: (20240613 - Murky) - need to use type guard instead
   const isCompanyIdValid = !Number.isNaN(Number(companyId));
   const isAccountIdValid = isParamNumeric(accountId);
